+++ poem/src/endpoint/mod.rs
@@ -4,12 +4,13 @@ mod after;
 mod and_then;
 mod around;
 mod before;
+mod catch_all_error;
 mod catch_error;
 #[allow(clippy::module_inception)]
 mod endpoint;
 mod files;
+mod inspect_all_err;
 mod inspect_err;
-mod inspect_typed_err;
 mod map;
 mod map_to_response;
 #[cfg(feature = "prometheus")]
@@ -21,11 +22,12 @@ pub use after::After;
 pub use and_then::AndThen;
 pub use around::Around;
 pub use before::Before;
+pub use catch_all_error::CatchAllError;
 pub use catch_error::CatchError;
 pub use endpoint::{make, make_sync, BoxEndpoint, Endpoint, EndpointExt, IntoEndpoint};
 pub use files::Files;
+pub use inspect_all_err::InspectAllError;
 pub use inspect_err::InspectError;
-pub use inspect_typed_err::InspectTypedError;
 pub use map::Map;
 pub use map_to_response::MapToResponse;
 #[cfg(feature = "prometheus")]