+++ poem/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# Unreleased
+
+- Add `Endpoint::get_response` method.
+
 # [1.2.2] 2021-12-16
 
 - Add `EndpointExt::inspect_typed_err` method.